+++ example/www/main.js
@@ -287,7 +287,7 @@ webpackContext.id = "../node_modules/moment/locale sync recursive ^\\.\\/.*$";
 /*! exports provided: name, version, description, main, typings, scripts, repository, keywords, author, license, bugs, homepage, devDependencies, private, default */
 /***/ (function(module) {
 
-module.exports = {"name":"ack-angular","version":"1.11.6","description":"Extra special directives, components, providers and pipes to aide in tackling everyday interface development needs in Angular2","main":"dist/index.js","typings":"dist/index.d.ts","scripts":{"start":"ack-reload -d example/www","build:index":"pug example/src/index.pug --out example/src","build:css":"npm-run-all build:css:ack-css-boot build:css:csshake build:css:dist build:css:example","build:css:ack-css-boot":"ack-sass node_modules/ack-css-boot/scss/ack-css-boot.scss dist/ack-css-boot.css --production","build:css:csshake":"ack-sass node_modules/csshake/scss/csshake.scss dist/csshake.css --production","build:css:dist":"ack-sass scss/ack-angular.scss dist/ack-angular.css --production","build:css:example":"ack-sass scss/ack-angular.scss example/src/assets/styles.css --production","install:example":"npm install --prefix example","test":"ng test --browser PhantomJS --single-run","test:watch":"ng test","build:dist":"ngc --declaration --project src","watch:dist":"watch \"npm-run-all build:dist\" src/ --ignoreDirectoryPattern=/pugs/","build:assets":"npm-run-all build:assets:src build:assets:example","build:assets:src":"ack-pug-bundler src/components/pugs/ src/components/templates/ --oneToOne --outType ts","build:assets:example":"ack-pug-bundler example/src/pugs/ example/src/templates/ --oneToOne --outType ts","watch:assets":"npm-run-all --parallel \"build:assets:src -- --watch\" \"build:assets:example -- --watch\"","watch":"npm-run-all --parallel build:index watch:dist watch:assets watch:js","watch:js":"ng serve example --port 4201 --open","build:js":"ng build example","build:universal":"npm-run-all build:index build:css build:assets compile:dist:package","build":"npm-run-all build:universal build:dist build:js","compile:dist:package":"node scripts/update-dist-package.js"},"repository":{"type":"git","url":"git+https://github.com/AckerApple/ack-angular.git"},"keywords":["ng","Angular","directives","components","pipes","providers"],"author":"Acker Dawn Apple","license":"MIT","bugs":{"url":"https://github.com/AckerApple/ack-angular/issues"},"homepage":"https://github.com/AckerApple/ack-angular#readme","devDependencies":{"@angular-devkit/build-angular":"~0.7.0","@angular/animations":"^6.1.1","@angular/cli":"^6.1.2","@angular/common":"^6.1.1","@angular/compiler":"^6.1.1","@angular/compiler-cli":"^6.1.1","@angular/core":"^6.1.1","@angular/forms":"^6.1.1","@angular/platform-browser":"^6.1.1","@angular/platform-browser-dynamic":"^6.1.1","@angular/router":"^6.1.1","@types/jasmine":"^2.8.6","ack-angular-fx":"^3.0.0","ack-css-boot":"^1.2.53","ack-pug-bundler":"^1.3.15","ack-reload":"^2.0.12","ack-sass":"^1.1.1","ack-x":"^1.4.5","animate.css":"^3.7.0","classlist-polyfill":"^1.2.0","csshake":"^1.5.3","jasmine":"^3.1.0","karma":"^2.0.5","karma-chrome-launcher":"^2.2.0","karma-coverage-istanbul-reporter":"^2.0.1","karma-jasmine":"^1.1.2","karma-jasmine-html-reporter":"^1.2.0","karma-phantomjs-launcher":"^1.0.4","localforage":"^1.7.2","ngx-page-scroll":"^5.0.0","npm-run-all":"^4.1.3","phantomjs-prebuilt":"^2.1.16","pug":"^2.0.1","pug-cli":"^1.0.0-alpha6","reflect-metadata":"^0.1.12","rxjs":"^6.4.0","ts-helpers":"^1.1.2","ts-loader":"^4.0.1","ts-node":"^5.0.1","typescript":"2.9.2","zone.js":"^0.8.26"},"private":true};
+module.exports = {"name":"ack-angular","version":"1.11.7","description":"Extra special directives, components, providers and pipes to aide in tackling everyday interface development needs in Angular2","main":"dist/index.js","typings":"dist/index.d.ts","scripts":{"start":"ack-reload -d example/www","build:index":"pug example/src/index.pug --out example/src","build:css":"npm-run-all build:css:ack-css-boot build:css:csshake build:css:dist build:css:example","build:css:ack-css-boot":"ack-sass node_modules/ack-css-boot/scss/ack-css-boot.scss dist/ack-css-boot.css --production","build:css:csshake":"ack-sass node_modules/csshake/scss/csshake.scss dist/csshake.css --production","build:css:dist":"ack-sass scss/ack-angular.scss dist/ack-angular.css --production","build:css:example":"ack-sass scss/ack-angular.scss example/src/assets/styles.css --production","install:example":"npm install --prefix example","test":"ng test --browser PhantomJS --single-run","test:watch":"ng test","build:dist":"ngc --declaration --project src","watch:dist":"watch \"npm-run-all build:dist\" src/ --ignoreDirectoryPattern=/pugs/","build:assets":"npm-run-all build:assets:src build:assets:example","build:assets:src":"ack-pug-bundler src/components/pugs/ src/components/templates/ --oneToOne --outType ts","build:assets:example":"ack-pug-bundler example/src/pugs/ example/src/templates/ --oneToOne --outType ts","watch:assets":"npm-run-all --parallel \"build:assets:src -- --watch\" \"build:assets:example -- --watch\"","watch":"npm-run-all --parallel build:index watch:dist watch:assets watch:js","watch:js":"ng serve example --port 4201 --open","build:js":"ng build example","build:universal":"npm-run-all build:index build:css build:assets compile:dist:package","build":"npm-run-all build:universal build:dist build:js","compile:dist:package":"node scripts/update-dist-package.js"},"repository":{"type":"git","url":"git+https://github.com/AckerApple/ack-angular.git"},"keywords":["ng","Angular","directives","components","pipes","providers"],"author":"Acker Dawn Apple","license":"MIT","bugs":{"url":"https://github.com/AckerApple/ack-angular/issues"},"homepage":"https://github.com/AckerApple/ack-angular#readme","devDependencies":{"@angular-devkit/build-angular":"~0.7.0","@angular/animations":"^6.1.1","@angular/cli":"^6.1.2","@angular/common":"^6.1.1","@angular/compiler":"^6.1.1","@angular/compiler-cli":"^6.1.1","@angular/core":"^6.1.1","@angular/forms":"^6.1.1","@angular/platform-browser":"^6.1.1","@angular/platform-browser-dynamic":"^6.1.1","@angular/router":"^6.1.1","@types/jasmine":"^2.8.6","ack-angular-fx":"^3.0.0","ack-css-boot":"^1.2.53","ack-pug-bundler":"^1.3.15","ack-reload":"^2.0.12","ack-sass":"^1.1.1","ack-x":"^1.4.5","animate.css":"^3.7.0","classlist-polyfill":"^1.2.0","csshake":"^1.5.3","jasmine":"^3.1.0","karma":"^2.0.5","karma-chrome-launcher":"^2.2.0","karma-coverage-istanbul-reporter":"^2.0.1","karma-jasmine":"^1.1.2","karma-jasmine-html-reporter":"^1.2.0","karma-phantomjs-launcher":"^1.0.4","localforage":"^1.7.2","ngx-page-scroll":"^5.0.0","npm-run-all":"^4.1.3","phantomjs-prebuilt":"^2.1.16","pug":"^2.0.1","pug-cli":"^1.0.0-alpha6","reflect-metadata":"^0.1.12","rxjs":"^6.4.0","ts-helpers":"^1.1.2","ts-loader":"^4.0.1","ts-node":"^5.0.1","typescript":"2.9.2","zone.js":"^0.8.26"},"private":true};
 
 /***/ }),
 
@@ -2503,7 +2503,7 @@ var AckArray = /** @class */ (function () {
         }
         this.inited = true;
         Promise.resolve().then(function () {
-            return _this.loop();
+            return _this.loop(true);
         });
     };
     //watch array deep changes
@@ -2514,7 +2514,7 @@ var AckArray = /** @class */ (function () {
         var changes = this.iterableDiffer.diff(this.array);
         if (changes) {
             Promise.resolve().then(function () {
-                return _this.loop();
+                return _this.loop(false);
             });
         }
     };
@@ -2527,7 +2527,7 @@ var AckArray = /** @class */ (function () {
         }
         if (this.inited && loop) {
             Promise.resolve().then(function () {
-                return _this.loop();
+                return _this.loop(true);
             });
         }
     };
@@ -2562,11 +2562,11 @@ var AckArray = /** @class */ (function () {
         }
         return item;
     };
-    AckArray.prototype.loop = function () {
+    AckArray.prototype.loop = function (reset) {
         if (!this.array) {
             this.array = [];
         }
-        this.loopStart.emit();
+        this.loopStart.emit(reset);
         var last = this.array.length;
         for (var x = 0; x < last; ++x) {
             this.loopEach.emit({ index: x, item: this.array[x] });
@@ -2593,10 +2593,12 @@ var AckArray = /** @class */ (function () {
         this.pushed.createPages = true;
         var pos = 0;
         var last = 0;
-        this.loopStart.subscribe(function () {
+        this.loopStart.subscribe(function (reset) {
             pos = 0;
             last = _this.array.length;
-            _this.pageChange.emit(_this.page = 0);
+            if (reset) {
+                _this.pageChange.emit(_this.page = 0);
+            }
             _this.pages = _this.pages || [];
             _this.pages.length = 0; //dont break binding if pages remembered
             _this.pages.push([]);
@@ -2609,6 +2611,9 @@ var AckArray = /** @class */ (function () {
             }
         });
         this.loopEnd.subscribe(function () {
+            if (_this.page && _this.page >= _this.pages.length) {
+                _this.pageChange.emit(_this.page = 0);
+            }
             _this.pagesChange.emit(_this.pages);
         });
     };
@@ -2617,7 +2622,7 @@ var AckArray = /** @class */ (function () {
         this.array.length = 0;
         this.array.push(item);
         this.arrayChange.emit(this.array);
-        this.loop();
+        this.loop(true);
     };
     //looks up id or the item itself is an ID
     AckArray.prototype.getItemId = function (item, itemIndexName) {
@@ -2654,7 +2659,7 @@ var AckArray = /** @class */ (function () {
     };
     AckArray.prototype.push = function (item) {
         this.param().push(item);
-        this.loop();
+        this.loop(false);
         return this;
     };
     AckArray.prototype.unshift = function (item) {
@@ -2664,7 +2669,7 @@ var AckArray = /** @class */ (function () {
     AckArray.prototype.splice = function (x, y) {
         if (y === void 0) { y = 1; }
         this.param().splice(x, y);
-        this.loop();
+        this.loop(false);
         return this;
     };
     AckArray.prototype.param = function () {
@@ -2744,7 +2749,7 @@ var AckArray = /** @class */ (function () {
             this.sortArray.pop();
         }
         this.inSort = false;
-        this.loop(); //cause pages to be updated
+        this.loop(true); //cause pages to be updated
     };
     __decorate([
         core_1.Input(),